feat(5): converted domain types to zod schamas
Converted all domain types to zod schemas. Those schemas are then used
to infer the types for type script. The importanace of this is that the
zod schemas can be used with the elysia api to infer input params or to
make it easier to define type and schemas for elysia arguments and the
like.

Linked to: #5